pyproject.toml
[tool.poetry]
name = "montecarlonfl"
version = "0.1.0"
description = ""
authors = ["Chris Aberle <35345242+chrisaberle@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.3"
numpy = "^1.25.2"
polars = "^0.18.15"
pyarrow = "^13.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
requests-html = "^0.10.0"
websockets = "8.1"
playwright = "^1.39.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
